Rename local SDMX support package and remove facet extraction from metadata extractor
