+++ tools/agentic_import/sdmx_support/metadata_enricher_fetch_test.py
@@ -15,6 +15,7 @@
 # limitations under the License.
 
 import json
+import os
 import tempfile
 import unittest
 from pathlib import Path
@@ -22,20 +23,20 @@ from unittest import mock
 
 from jinja2 import Template
 
-from tools.agentic_import.sdmx.metadata_enricher_find import (
-    Config, EnrichmentItemsFinder)
+from tools.agentic_import.sdmx_support.metadata_enricher_fetch import (
+    Config, EnrichmentDataFetcher)
 
 
-class EnrichmentItemsFinderTest(unittest.TestCase):
+class EnrichmentDataFetcherTest(unittest.TestCase):
 
     def test_dry_run_creates_prompt_and_run_dir(self) -> None:
         with tempfile.TemporaryDirectory() as tmpdir:
-            input_path = Path(tmpdir) / 'metadata.json'
-            input_path.write_text(json.dumps({"dataflows": []}))
-            output_path = Path(tmpdir) / 'out' / 'items.json'
+            input_path = Path(tmpdir) / 'items.json'
+            input_path.write_text(json.dumps({"items": []}))
+            output_path = Path(tmpdir) / 'out' / 'items_enriched.json'
 
             config = Config(
-                input_metadata_json=str(input_path),
+                input_items_json=str(input_path),
                 dataset_prefix='demo',
                 output_path=str(output_path),
                 dry_run=True,
@@ -44,11 +45,11 @@ class EnrichmentItemsFinderTest(unittest.TestCase):
                 working_dir=tmpdir,
             )
 
-            finder = EnrichmentItemsFinder(config)
+            fetcher = EnrichmentDataFetcher(config)
             with mock.patch("jinja2.environment.Template.render",
                             autospec=True,
                             side_effect=Template.render) as render_mock:
-                result = finder.find_items_to_enrich()
+                result = fetcher.fetch_enrichment_data()
 
             self.assertTrue(result.run_id.startswith('demo_gemini_'))
             self.assertTrue(result.run_dir.is_dir())
@@ -67,7 +68,7 @@ class EnrichmentItemsFinderTest(unittest.TestCase):
             _, render_kwargs = render_mock.call_args
             self.assertEqual(
                 render_kwargs, {
-                    "input_metadata_abs": str(input_path.resolve()),
+                    "input_items_abs": str(input_path.resolve()),
                     "output_path_abs": str(output_path.resolve()),
                 })
